Feature: Tab through slides
The intention is to improve accessibility in cases
where the slideshow has multiple slides per "view",
and it should be possible to tab through each
individual slide.

The reason this is needed is that the keyboard
navigation (left arrow, right arrow) only operates
on a "view" basis, so you cannot easily select
individual slides.

This feature adds two additional options;
- options.tabThroughSlides (boolean, default false):
   setup the slider so that tab + shift/tab will let
   you switch back and forth between the individual slides.
- options.dotsTabbable (boolean, default true):
   Disable tab navigation for the dots; useful if you
   don't want to force the user to tab through all the dots
   if they can tab through the individual slides anyways.